+++ tests/test_randprocs/test_markov/test_discrete/test_discrete_transition.py
